+++ plant_listener/devices/include/plantlistener/device/device.hpp
@@ -57,7 +57,7 @@ class Device {
    * @param port The port to be read.
    * @returns value of sensor between [min_value, max_value], or -1 on error.
    */
-  virtual uint64_t readPort(const uint8_t port) = 0;
+  virtual double readPort(const uint8_t port) = 0;
 
   virtual nlohmann::json dump();
 };
